Fix grep errors
grep was unable to handle files/directories with spaces in their names the way that I wrote it. This change fixes that.

This only affected grep (logging to console). Previous runs of this script would have still worked as the stream engine was not affected by this. Users do not need to re-run this script after this change!!